pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vadim Shmakov <v@shmakov.me>"]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1"
litestar = { extras = ["standard"], version = "^2" }
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.3"
black = "^24.4.2"
mypy = "^1.10.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
